requirements.txt
# Direct dependencies
Django==3.0.9
django-extensions==3.0.3
django-ratelimit==3.0.1
django-redis==4.12.1
djangorestframework==3.11.0
factory-boy==2.12.0
gunicorn==20.0.4
gunicorn-color==0.1.0
ipython==7.17.0
msgpack==1.0.0
parameterized==0.7.4
python-rapidjson==0.9.1
psycopg2-binary==2.8.5
sentry-sdk==0.16.3
django-allauth==0.42.0
django-annoying==0.10.6

# Indirect dependencies
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
click==7.1.2
decorator==4.4.2
Faker==4.1.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.10
wait-for-it==2.0.1
wcwidth==0.2.5
requests-oauthlib==1.3.0